ENH: break PlayerProfile into smaller components
