+++ src/app/components/PlayerProfile/MostPlayed/styled.js
@@ -1,0 +1,13 @@
+import styled from 'styled-components'
+
+import CoreImage from 'app/components/core/Image'
+
+export const Container = styled.div`
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+`
+
+export const Hero = styled(CoreImage)`
+  width: 50px;
+`
