+++ src/app/components/PlayerProfile/Rank/styled.js
@@ -1,0 +1,14 @@
+import styled from 'styled-components'
+
+import CoreImage from 'app/components/core/Image'
+
+export const Container = styled.div``
+
+export const Star = styled(CoreImage)`
+  width: 150px;
+  position: absolute;
+`
+
+export const Medal = styled(CoreImage)`
+  width: 150px;
+`
